wip: fan out lambda-layer publishing in CI?
Kind of neat to make Circle do all the publishing in parallel. Not 100%
sure it's _better_.

It takes about 1 minute for the aws-cli orb to spin up and configure
itself. Then it takes about 2 minutes total to publish the extension
to all listed regions in a loop.

Possible advantage to publishing fan out:
* slightly faster because parallelism?
* look up a region's ARN by looking at the output in the job with a name
  for that region

Possible disadvantage to fan out:
* not _really_ that much faster
* all the resultant ARNs in one job's output makes it easy to scan and
  check that the regions have a consistent layer version at the end