parse JSON emitted to stdout from libhoney transmissions correctly (#21)
* update msg parsing to look for libhoney data

When it's JSON and the 'data' key is present, that's probably an event
from a libhoney transmission sending to stdout. If 'data' is a map, use
it for the event's fields. Otherwise, treat the the message as flat JSON
and use all its keys as fields for the event.

Also updates the variable names used for type checks to make it
clearer what they are when they're used. 'v' is vague when there are
multiple type checks occurring in the same context.

* update time parsing to detect a libhoney timestamp

JSON emitted by a libhoney's transmission will have a 'time' key for the
timestamp. Prefer that, then go on to try other ways to determine time.