move zipping from publish script to makefile
Make is good at knowing how to make files. This moves the layer contents
zip creation to make. There's some duplication, but I think that keeps
the Making more readable.

Publish script updated to check for zip existence instead of creating
them. A little layer_name declaration refactor in preparation for
version-specific layer names in the future.

CI updated to use the `zip` make target to build the binaries and the
zip files for artifact storage and pass-along.