update list of regions that support multi-arch
The combination of all of the regions we've been publishing to matches
the list of regions that support both arm64 and x86_64 at the AWS Lambda
pricing page as of 2022/10/12. Moving existing regions to the WITH_ARCH
list.

NO_ARCH is currently empty, but one new-to-us region is online and
currently only supports x86_64. We can add it to NO_ARCH when we've
granted the publishing credentials permission to publish layers there.