maint: release process updates for friendlier ARNs (#103)
## Which problem is this PR solving?

Release woes, enumerated below with the changes.

## ~Short~ description of the changes

### Reorganize releasing

- Closes #82 
- Rename published layer in regions to
`honeycomb-lambda-extension-<arch>-<version tag>` so that our layer
versions will always be `:1`
- Some changes in service towards #95, but not fully producing a
release.json yet

Make is good at knowing how to _make_ files. This moves the layer
contents
zip creation from the publish script to the Makefile. There's some
duplication,
but I think that keeps the Making more readable. CI updated to use the
`zip`
make target to build the binaries _and_ the zip files for artifact
storage and
workflow pass-along.

Publish script updated to record the results of AWS API calls to files
instead of
to script variables or the console. With the results recorded on the
filesystem,
we can see success/failures after the fact and thereby build
release.json from
the data in the API responses.

### Use the CIRCLE_TAG as the version for a build

- Closes #66 

Include the build version in the user-agent (formatting updated to
spec).

If the CI workflow doesn't have a tag, default to an informative version
derived from the current state of the git repository. For example, this
git describe command returns the following from the branch this commit
will land on:

    » git describe --always --tags --match "v[0-9]*" HEAD
    v10.3.0-9-g834dbcb

Which is:
* the nearest tag in history that matches the pattern
* the number of commits the branch's HEAD is away from that tag/commit
* g + the sha of current HEAD


### Test against only a single Go version

- Closes #100 

This is a project released and consumed as a binary. We only need to
test
against the version of Go we build and release with.

Co-authored-by: Jamie Danielson <jamieedanielson@gmail.com>