feat: configurable HTTP transport connect timeout (#111)
## Which problem is this PR solving?

* Partially address #101 

This provides a means to configure just the connect timeout, probably
the most important of the timeout settings, so limiting scope to just
that.

## Short description of the changes

* A new env variable HONEYCOMB_CONNECT_TIMEOUT can be used to customize
the HTTP transport connection timeout for the underlying libhoney-go
client so that when there are network connectivity issues, batch upload
will fail faster and not hang for the full duration of the batch upload
timeout. The default connection timeout is set to 3 seconds if not
specified. This closes out a portion of issue #101, as it's only
handling connect timeout for now. Limiting network connection timeouts
is critical for a Lambda extension as it's possible for an extension to
extend the duration of a function invocation and result in function
timeouts. This has real world impacts beyond just costs. This scenario
can easily result in overall Lambda account concurrency limits being
exceeded, which will prevent further function executions from running
(essentially a denial of service). Furthermore, when function timeouts
occur, this causes the underyling Lambda sandbox to terminate resulting
in performance impact in the form of cold starts for all invocations.
* Refactor: move libhoney-go code out of main package and into an
eventpublisher package which is just a wrapper around the libhoney
client with reduced API surface that is configured with default network
settings more appropriate for a Lambda extension. This move to a
separate package also helps with the unit tests as this can be written
outside of package main.

Co-authored-by: Mike Goldsmith <goldsmith.mike@gmail.com>