maint: Update AWS regions we publish to (#123)
## Which problem is this PR solving?

Updates the list of AWS regions we publish the lambda extension to. Also
adds docs to how we source the list of regions and how we determine
which platforms it supports.

- Closes #121 

## Short description of the changes

- Update `publish.sh` with new regions
- Add section to README on how to get list of regions including
supported platforms
- Rename from `REGIONS_NO_ARCH` and `REGIONS_WITH_ARCH` to
`REGIONS_NO_ARM` and `REGIONS-WITH_ARM` for better clarity

## How to verify this has the expected result

- We published a dev tag for `ap-southeast-4` and `us-east-1` to confirm
the layer appeared as expected in both regions
(`honeycomb-lambda-extension-x86_64-v11-1-1-dev` in both,
`honeycomb-lambda-extension-arm64-v11-1-1-dev` in `us-east-1`). Also
clicked into each of the other new regions to confirm an empty layer
page, so we have permissions to publish there.

---------

Co-authored-by: JamieDanielson <jamieedanielson@gmail.com>
Co-authored-by: Purvi Kanal <purvikanal@honeycomb.io>