fix(eventprocessor): delay shutdown to allow lambda app to flush (#162)
## Which problem is this PR solving?

- Events are missing from the Lambda application instrumented with the
extension, because the extension shutdown races the application's own
shutdown/flush procedures. Per
https://docs.aws.amazon.com/lambda/latest/dg/runtimes-extensions-api.html#runtimes-lifecycle-extensions-shutdown,
it can take up to 500ms for an internal extension using the sigterm
handler to shut down, and we're given the shutdown for each process in
parallel but our work shouldn't take the full 2k seconds.
- Fixes #141

## Short description of the changes

- Delays until 100ms before the deadline before triggering flush and
shutdown, to allow for telemetry from the underlying application to be
sent.

---------

Co-authored-by: Mike Goldsmith <goldsmith.mike@gmail.com>